+++ app/assets/config/manifest.js
@@ -1,6 +1,5 @@
 //= link_tree ../images
 //= link_directory ../stylesheets .css
-//= link_tree ../javascript .js
 //
 //= link govuk-frontend/govuk/all.js
 //= link govuk-frontend/govuk/assets/images/favicon.ico
@@ -11,4 +10,7 @@
 //= link govuk-frontend/govuk/assets/images/govuk-apple-touch-icon.png
 //= link govuk-frontend/govuk/assets/images/govuk-opengraph-image.png
 //= link "@ministryofjustice/frontend/moj/assets/images/moj-logotype-crest.png"
+//= link dfefrontend.js
 //= link accessible-autocomplete/dist/accessible-autocomplete.min.js
+
+//= link_tree ../builds