+++ app/policies/assignment_policy.rb
@@ -5,6 +5,16 @@ class AssignmentPolicy
     @user = user
   end
 
+  def edit_assigned_user?
+    return true if @user.is_service_support?
+
+    @user.has_role?
+  end
+
+  def update_assigned_user?
+    edit_assigned_user?
+  end
+
   def assign_team_leader?
     return true if @user.is_service_support?
 
@@ -23,16 +33,6 @@ class AssignmentPolicy
 
   def update_regional_delivery_officer?
     assign_regional_delivery_officer?
-  end
-
-  def assign_assigned_to?
-    return true if @user.is_service_support?
-
-    @user.has_role?
-  end
-
-  def update_assigned_to?
-    assign_assigned_to?
   end
 
   def assign_team?