+++ config/routes.rb
@@ -29,10 +29,12 @@ Rails.application.routes.draw do
       post "team-lead", action: :update_team_leader
       get "regional-delivery-officer", action: :assign_regional_delivery_officer
       post "regional-delivery-officer", action: :update_regional_delivery_officer
-      get "assigned-to", action: :assign_assigned_to
-      post "assigned-to", action: :update_assigned_to
       get "team", action: :assign_team
       post "team", action: :update_team
+    end
+    namespace :internal_contacts, path: "internal-contacts" do
+      get "assigned-user/edit", action: :edit_assigned_user
+      put "assigned-user/", action: :update_assigned_user
     end
   end
 
@@ -284,6 +286,7 @@ Rails.application.routes.draw do
 
   # Search
   get "search", to: "search#results"
+  get "search/user/", to: "search#user"
 
   # Defines the root path route ("/")
   root "root#home"